Stopping panning for the axis individually
I’ve enhanced the performChange method on the BarLineChartViewBase to
return the status of panning for each axis.
Previously this method just return whether any of the two axis is still
moving. Now it returns a tuple of two booleans, representing the
movement state of the X- and YAxis.

Additional I change the decelerationLoop method implementation, to set
the deceleration velocity’s x and y values individually to zero,
depending on the return value of the perfromPanChange method.

All this is necessary to prevent the further deceleration of the XAxis
if it reached the edge and then a new dataset is added, while the YAxis
is still moving.